agent/requirements.txt
requests==2.31.0
psutil==5.9.8
pyyaml==6.0.1
